[containers] GeneralizedTrie: Wshorten fixes

+++ include/sserialize/containers/GeneralizedTrie/SerializableTrie.h
@@ -97,7 +97,7 @@ checkTrieEqualityRecurse(Node * curNode, Static::TrieNode curStaticNode, std::st
 		return false;
 	}
 
-	for(size_t i = 0; i < curStaticNode.childCount(); i++) {
+	for(uint32_t i = 0; i < curStaticNode.childCount(); i++) {
 		uint32_t childChar = curStaticNode.childCharAt(i);
 		std::string childCharStr = "";
 		utf8::append(childChar, std::back_inserter(childCharStr));
@@ -171,7 +171,7 @@ checkIndexEqualityRecurse(Node* curNode, sserialize::Static::TrieNode curStaticN
 		}
 	}
 
-	for(size_t i = 0; i < curStaticNode.childCount(); i++) {
+	for(uint32_t i = 0; i < curStaticNode.childCount(); i++) {
 		uint32_t childChar = curStaticNode.childCharAt(i);
 		if (curNode->children.count(childChar) == 0) {
 			return false;
